feat: framework-independent `RootStyleRegistry` implementation
