+++ .changeset/stale-lemons-wonder.md
@@ -1,0 +1,5 @@
+---
+"@makeswift/runtime": patch
+---
+
+feat: framework-independent `RootStyleRegistry` implementation
