+++ packages/runtime/src/next/root-style-registry.tsx
@@ -1,38 +1,14 @@
 'use client'
 
-import createCache, { EmotionCache } from '@emotion/cache'
-import { cache } from '@emotion/css'
+import { type ReactNode, useState } from 'react'
 import { useServerInsertedHTML } from 'next/navigation'
-import { ReactNode, createContext, useContext, useState } from 'react'
 
-const CacheContext = createContext(cache)
+import {
+  createRootStyleCache,
+  RootStyleRegistry as ReactRootStyleRegistry,
+} from '../runtimes/react/root-style-registry'
 
-const DEFAULT_CSS_RESET_ENABLED = true
-const CSSResetEnabledContext = createContext(DEFAULT_CSS_RESET_ENABLED)
-
-const createRootStyleCache = ({ key }: { key: string }) => {
-  const cache = createCache({ key })
-  cache.compat = true
-
-  const prevInsert = cache.insert
-  let inserted: string[] = []
-
-  cache.insert = (...args) => {
-    const serialized = args[1]
-    if (cache.inserted[serialized.name] === undefined) {
-      inserted.push(serialized.name)
-    }
-    return prevInsert(...args)
-  }
-
-  const flush = () => {
-    const prevInserted = inserted
-    inserted = []
-    return prevInserted
-  }
-
-  return { cache, flush }
-}
+export { useCache, useCSSResetEnabled } from '../runtimes/react/root-style-registry'
 
 type Props = {
   children: ReactNode
@@ -44,12 +20,8 @@ type Props = {
   enableCssReset?: boolean
 }
 
-export function RootStyleRegistry({
-  children,
-  cacheKey,
-  enableCssReset = DEFAULT_CSS_RESET_ENABLED,
-}: Props) {
-  const [{ cache, flush }] = useState(() => createRootStyleCache({ key: cacheKey ?? 'mswft' }))
+export function NextRootStyleRegistry({ children, cacheKey, enableCssReset }: Props) {
+  const [{ cache, flush }] = useState(() => createRootStyleCache({ key: cacheKey }))
 
   useServerInsertedHTML(() => {
     const names = flush()
@@ -70,18 +42,8 @@ export function RootStyleRegistry({
   })
 
   return (
-    <CacheContext.Provider value={cache}>
-      <CSSResetEnabledContext.Provider value={enableCssReset}>
-        {children}
-      </CSSResetEnabledContext.Provider>
-    </CacheContext.Provider>
+    <ReactRootStyleRegistry cache={cache} enableCssReset={enableCssReset}>
+      {children}
+    </ReactRootStyleRegistry>
   )
-}
-
-export function useCache(): EmotionCache {
-  return useContext(CacheContext)
-}
-
-export function useCSSResetEnabled(): boolean {
-  return useContext(CSSResetEnabledContext)
 }
